Change x86 code from crate into module
