+++ Cargo.toml
@@ -8,7 +8,29 @@ authors = [ "robert.w.gries@gmail.com" ]
 crate-type = ["staticlib"]
 
 [target.'cfg(any(target_arch = "x86", target_arch = "x86_64"))'.dependencies]
-arch_x86 = { path = "arch/x86", features = ["serial"] }
+bit_field = "0.7.0"
+bitflags = "0.9.1"
 # TODO: Submit issue for multiboot2-elf64 repo
 # multiboot2 = { git = "https://github.com/phil-opp/multiboot2-elf64", features = [] }
 multiboot2 = { git = "https://github.com/ahmedcharles/multiboot2-elf64", branch = "elf32" }
+once = "0.3.2"
+rlibc = "1.0.0"
+spin = "0.4.5"
+volatile = "0.1.0"
+
+[dependencies.x86]
+git = "https://github.com/gz/rust-x86"
+branch = "master"
+default-features = false
+
+[dependencies.lazy_static]
+version = "0.2.4"
+features = ["spin_no_std"]
+
+[dependencies.hole_list_allocator]
+path = "allocator/hole_list_allocator"
+
+[features]
+default = ["serial"]
+serial = []
+vga = []
