+++ Makefile
@@ -40,13 +40,13 @@ rust_target ?= $(rust_arch)-rxinu
 rust_os := target/$(rust_target)/debug/librxinu.a
 
 # Source files
-linker_script := arch/$(arch)/asm/linker.ld
-grub_cfg := arch/$(arch)/asm/grub.cfg
-ASM_SRC := $(wildcard arch/$(arch)/asm/*.nasm) \
-	$(wildcard arch/$(arch)/asm/$(target)/*.nasm)
+linker_script := src/arch/$(arch)/asm/linker.ld
+grub_cfg := src/arch/$(arch)/asm/grub.cfg
+ASM_SRC := $(wildcard src/arch/$(arch)/asm/*.nasm) \
+	$(wildcard src/arch/$(arch)/asm/$(target)/*.nasm)
 
 # Object files
-ASM_OBJ := $(patsubst arch/$(arch)/asm/%.nasm, build/arch/$(arch)/asm/%.o, $(ASM_SRC))
+ASM_OBJ := $(patsubst src/arch/$(arch)/asm/%.nasm, build/arch/$(arch)/asm/%.o, $(ASM_SRC))
 
 .PHONY: all cargo clean debug gdb iso run
 
@@ -83,7 +83,7 @@ $(kernel): cargo $(rust_os) $(ASM_OBJ) $(linker_script)
 	@$(LD) $(LDFLAGS) -T $(linker_script) -o $(kernel) $(ASM_OBJ) $(rust_os)
 
 # compile architecture specific files
-build/arch/$(arch)/asm/%.o: arch/$(arch)/asm/%.nasm
+build/arch/$(arch)/asm/%.o: src/arch/$(arch)/asm/%.nasm
 	@mkdir -p $(shell dirname $@)
 	@echo "  Assembling $<"
 	@$(ASM) $(ASFLAGS) $< -o $@
