--- arch/x86/Cargo.toml
@@ -1,31 +1,0 @@
-[package]
-name = "arch_x86"
-version = "0.1.0"
-authors = [ "robert.w.gries@gmail.com" ]
-
-[dependencies]
-bit_field = "0.7.0"
-bitflags = "0.9.1"
-# TODO: Submit issue for broken multiboot2-elf64
-# multiboot2 = { git = "https://github.com/phil-opp/multiboot2-elf64", features = []}
-multiboot2 = { git = "https://github.com/ahmedcharles/multiboot2-elf64", branch = "elf32" }
-once = "0.3.2"
-rlibc = "1.0.0"
-spin = "0.4.5"
-volatile = "0.1.0"
-
-[dependencies.x86]
-git = "https://github.com/gz/rust-x86"
-branch = "master"
-default-features = false
-
-[dependencies.lazy_static]
-version = "0.2.4"
-features = ["spin_no_std"]
-
-[dependencies.hole_list_allocator]
-path = "../../allocator/hole_list_allocator"
-
-[features]
-serial = []
-vga = []
