+++ src/arch/x86/asm/check-common.nasm
