+++ src/arch/x86/asm/startup-common.nasm
