+++ src/arch/x86/asm/x86_64/check_long_mode.nasm
