+++ src/arch/x86/device/mod.rs
@@ -1,4 +1,3 @@
-mod io;
 pub mod pic;
 pub mod serial;
 pub mod vga;
