+++ src/arch/x86/device/vga.rs
