+++ src/arch/x86/interrupts/irq.rs
@@ -1,10 +1,8 @@
-#![allow(unused_variables)]
+use arch::x86::device::pic;
+use arch::x86::device::serial::{COM1, COM2};
+use arch::x86::interrupts::idt::ExceptionStackFrame;
 
-use interrupts::idt::ExceptionStackFrame;
-
-use device::pic;
-use device::serial::{COM1, COM2};
-
+#[allow(unused_variables)]
 pub extern "x86-interrupt" fn cascade(stack_frame: &mut ExceptionStackFrame) {
     pic::MASTER.lock().ack();
 }
